requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
PyJWT==2.8.0
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
wheel==0.43.0
